# Optimized `filterDOMProps`.